Refactor : 코드 리펙토링 수정
코드 리펙토링을 리뷰를 반영하여 수정하였습니다.
