+++ src/main/java/nbbang/com/nbbang/domain/party/controller/ManyPartyResponseMessage.java
@@ -1,4 +1,4 @@
-package nbbang.com.nbbang.domain.parties.controller;
+package nbbang.com.nbbang.domain.party.controller;
 
 public class ManyPartyResponseMessage {
     public static final String ILLEGAL_PLACE = "위치 정보를 올바르게 입력하세요.";
