+++ src/main/java/nbbang/com/nbbang/domain/party/service/ManyPartyService.java
@@ -1,4 +1,4 @@
-package nbbang.com.nbbang.domain.parties.service;
+package nbbang.com.nbbang.domain.party.service;
 
 import lombok.RequiredArgsConstructor;
 import nbbang.com.nbbang.domain.party.domain.Party;
