+++ src/main/java/nbbang/com/nbbang/global/validator/PartyMemberValidatorService.java
@@ -4,7 +4,7 @@ import lombok.RequiredArgsConstructor;
 import nbbang.com.nbbang.domain.member.domain.Member;
 import nbbang.com.nbbang.domain.member.service.MemberService;
 import nbbang.com.nbbang.domain.party.domain.Party;
-import nbbang.com.nbbang.domain.parties.service.PartyService;
+import nbbang.com.nbbang.domain.party.service.PartyService;
 import org.springframework.stereotype.Service;
 
 @Service
